Whoopsie! Corrected some parsing mistakes.

+++ ClientDefinitions/DBC/Cataclysm/NumTalentsAtLevelEntry.cs
@@ -3,6 +3,6 @@
     public sealed class NumTalentsAtLevelEntry
     {
         public uint Level;
-        public uint TalentPointCount;
+        public float TalentPointCount;
     }
 }
